Fix bdos emulation issues found when testing submit.
Implementing the seach for '$' in the filenames in the current directory at disk reset.

Having the BDOS open file function check for the file being open already in the stfps array. If it is then it does not open it again.

Having the close function truncate the file according to the RC value in the FCB but only if the S2/EX values are pointing to the last extent of the file.